functions/write-to-bigquery/requirements.txt
functions-framework==3.*
flask
google-cloud-bigquery
datetime